move the `intersect_bank_lines_with_mesh` to the `utils.BankLinesProcessor` class as `intersect_with_mesh`
